PR cleanup, also moved all scheduling extensions into one file
