+++ super_editor/lib/src/super_textfield/infrastructure/text_scrollview.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
-import 'package:super_editor/src/infrastructure/flutter/flutter_pipeline.dart';
+import 'package:super_editor/src/infrastructure/flutter/flutter_scheduler.dart';
 import 'package:super_text_layout/super_text_layout.dart';
 import 'package:super_editor/src/super_textfield/super_textfield.dart';
